requirements.txt
python-chess >= 1.999
jupyter
numpy
flask
